backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic-settings==2.3.0
httpx==0.27.0
python-multipart==0.0.9
Pillow>=10.4.0
timm>=1.0.3
transformers>=4.40.0
torch>=2.3.0
torchvision>=0.18.0
numpy>=1.26.0
einops>=0.7.0
datasets>=2.19.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
lightgbm>=4.3.0
pytest==8.2.0
pytest-asyncio==0.23.7
